10.5.1 - fixes bugs in the new, unlimited precision, integer arithmetic subsystem
bigint will take mal-formed integers:
   (bigint "hello") -> nil
   (bigint "123hello") -> 123L
   (bigint "+123hello") -> 123L
   (bigint "-123hello") -> -123L
   (bigint "123.456") -> 123L
but should not be used on string of scientific notation, it only converts the mantissa
   (bigint "123.456e10") -> 123L

Fixed division when aligned or not aligned embedded bigint 0s (9 zeros).

Fixed more big integer division problems and added more tests to qa-bigint.

Compiler optimizations are now supressed in part of the big integer
division code to make it work on Linux. Does not affect performance.