10.7.0 - adds new functionality to existing functions, adds one new function and fixes bugs.
10.7.0 version 10.6.5 released as stable release January 2016.

In the crypto.lsp module "cdecl" must be used when importing on Windows.
(since v.10.6.3 a new compiler setup is used)

A second PRNG with variable initial N for seed initialization
via (seed S true N) and inquiry of current seed state via (seed).
This PRNG is built-in with same results across all platforms and
compilers. The current PRNG will stay as it offers a bigger set of
random numbers on Unix platforms. This new PRNG is necessary for doing
computer simulations and AI art, which is reproducable across
platforms and where random number sequences can be broken up in
repeatable layers using the (seed) instruction to save state.
When using the default libc based PRNG (seed) now returns the last
seed set by the programmer. On the new PRNG (seed) returns the last
seed state. - Not yet documented, may replace old PRNG in the future.

A (ssq vector) sum of squares function has been added. The vector can
be either a list or array vector.

guiserver.jar version 1.65 now correctly forces updates of a gs:canvas
in a gs:scroll-pane.

When a deleted symbol was prefix in a dynamic symbol, memory got corrupted.
The dynamic symbol reference is now replaced with nil like other symbol
references.

A illegal symbol when trying to create a context could crash the system,
when referenced again after the error message.